requirements.txt
feed_utils==0.1.8
sqlalchemy
psycopg2
kafka-python
beautifulsoup4
slimit
requests
pandas
flask
flask_classy
